+++ src/backend/joanie/core/models/products.py
@@ -536,14 +536,26 @@ class Order(BaseModel):
         constraints = [
             models.UniqueConstraint(
                 fields=["course", "owner", "product"],
-                condition=~models.Q(state=enums.ORDER_STATE_CANCELED),
-                name="unique_owner_course_product_not_canceled",
+                condition=~models.Q(
+                    state__in=[
+                        enums.ORDER_STATE_CANCELED,
+                        enums.ORDER_STATE_REFUNDING,
+                        enums.ORDER_STATE_REFUNDED,
+                    ]
+                ),
+                name="unique_owner_course_product_not_canceled_or_not_in_refund_states",
                 violation_error_message="An order for this product and course already exists.",
             ),
             models.UniqueConstraint(
                 fields=["enrollment", "owner", "product"],
-                condition=~models.Q(state=enums.ORDER_STATE_CANCELED),
-                name="unique_owner_enrollment_product_not_canceled",
+                condition=~models.Q(
+                    state__in=[
+                        enums.ORDER_STATE_CANCELED,
+                        enums.ORDER_STATE_REFUNDING,
+                        enums.ORDER_STATE_REFUNDED,
+                    ]
+                ),
+                name="unique_owner_enrollment_product_not_canceled_or_not_in_refund_states",
                 violation_error_message="An order for this product and enrollment already exists.",
             ),
             models.CheckConstraint(