+++ CHANGELOG.md
@@ -8,10 +8,17 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Fixed
+
+- Allow a user to create a new order when the previous order already
+  exists with the triplet owner-course-product or owner-enrollment-product
+
 ## [2.13.0] - 2025-01-15
 
 ### Added
 
+- `owner` field on the CreditCard model has been deprecated
+  and replaced with a many-to-many `owners` field
 - Add order export to CSV in back office
 - Branded Unicamp degrees